mig tarifas

+++ custom_inplast/models/product_template.py
@@ -15,7 +15,7 @@ class ProductTemplate(models.Model):
     mig_tipo_producto = fields.Integer('mig_tipo_producto')
     mig_unidad_logistica = fields.Integer('mig_unidad_logistica')
     mig_tipo_embalaje = fields.Float('mig_tipo_embalaje')
-    mig_peso_embalaje = fields.Integer('mig_peso_embalaje')
+    mig_peso_embalaje = fields.Float('mig_peso_embalaje')
 
     mig_cod_familia = fields.Char('mig_cod_familia')
     mig_cod_subfamilia = fields.Char('mig_cod_subfamilia')
